requirements.txt
dbt-core==1.7.17
dbt-sqlserver==1.7.4
pandas==2.0.3
plotly==5.22.0
pymssql==2.3.0
pyodbc==5.0.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.23.1
SQLAlchemy==2.0.31
st-pages==0.5.0
streamlit==1.39.0
webdriver-manager==4.0.2
websocket-client==1.8.0
